Properly overwrite filenames
This patch is based on the source of shred from coreutils by Colin Plumb.
While shred is now licensed under GPLv3, the parts of this patch that were
derived from shred could easily be changed to a version of shred
(git commit cad884a) that was licensed under GPLv2 or later.

Note that this patch only keeps the filename from being readable in the
filesystem structure. For journaling fileystems, such as ext{3,4}, the
filename is still retained in the journal.